refactor: replace with new global settings
