--- sbsp_remote_app/src/settings.rs
@@ -1,234 +1,0 @@
-// SPDX-License-Identifier: Elastic-2.0
-// Copyright (c) 2025 Keinsleif (https://github.com/Keinsleif)
-
-pub mod hotkey;
-pub mod manager;
-
-use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
-use ts_rs::TS;
-
-use hotkey::HotkeySettings;
-use sbsp_backend::model::cue::{
-    Cue, CueChain, CueColor, CueParam, FadeCueParam, LoadCueParam, PauseCueParam, StartCueParam,
-    StopCueParam, Uuid, WaitCueParam,
-    audio::{AudioCueParam, Decibels, Easing, FadeParam, SoundType},
-    group::{GroupCueParamBase, GroupMode},
-};
-
-#[derive(Serialize, Deserialize, Debug, Clone, Default, PartialEq, TS)]
-#[ts(export)]
-#[serde(rename_all = "camelCase", default)]
-pub struct GlobalRemoteSettings {
-    pub general: GeneralSettings,
-    pub appearance: AppearanceSettings,
-    pub hotkey: HotkeySettings,
-    pub template: TemplateSettings,
-    pub name_format: NameFormatSettings,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct GeneralSettings {
-    #[serde(alias = "advanceCursorWhenGo")]
-    pub advance_cursor_when_execute: bool,
-    pub lock_cursor_to_selection: bool,
-    pub copy_assets_when_add: bool,
-    pub seek_amount: f64,
-}
-
-impl Default for GeneralSettings {
-    fn default() -> Self {
-        Self {
-            advance_cursor_when_execute: false,
-            lock_cursor_to_selection: true,
-            copy_assets_when_add: false,
-            seek_amount: 5.0,
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct AppearanceSettings {
-    pub language: Option<String>,
-    pub dark_mode: DarkMode,
-    pub hide_controls: bool,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default, TS)]
-#[serde(rename_all = "camelCase")]
-pub enum DarkMode {
-    #[default]
-    Dark,
-    Light,
-    System,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct TemplateSettings {
-    pub audio: Cue,
-    pub wait: Cue,
-    pub fade: Cue,
-    pub start: Cue,
-    pub stop: Cue,
-    pub pause: Cue,
-    pub load: Cue,
-    pub group: Cue,
-}
-
-impl Default for TemplateSettings {
-    fn default() -> Self {
-        Self {
-            audio: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Audio(AudioCueParam {
-                    target: PathBuf::new(),
-                    start_time: None,
-                    fade_in_param: None,
-                    end_time: None,
-                    fade_out_param: None,
-                    volume: Decibels::IDENTITY,
-                    pan: 0.0,
-                    repeat: false,
-                    sound_type: SoundType::Streaming,
-                    envelope: Vec::new(),
-                }),
-            },
-            wait: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Wait(WaitCueParam { duration: 5.0 }),
-            },
-            fade: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Fade(FadeCueParam {
-                    target: Uuid::nil(),
-                    volume: Decibels::IDENTITY,
-                    fade_param: FadeParam {
-                        duration: 3.0,
-                        easing: Easing::InOutPow(2.0),
-                    },
-                }),
-            },
-            start: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Start(StartCueParam {
-                    target: Uuid::nil(),
-                }),
-            },
-            stop: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Stop(StopCueParam {
-                    target: Uuid::nil(),
-                    hard: false,
-                }),
-            },
-            pause: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Pause(PauseCueParam {
-                    target: Uuid::nil(),
-                }),
-            },
-            load: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Load(LoadCueParam {
-                    target: Uuid::nil(),
-                }),
-            },
-            group: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Group {
-                    base: GroupCueParamBase {
-                        mode: GroupMode::Playlist { repeat: true },
-                    },
-                    children: Vec::new(),
-                },
-            },
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct NameFormatSettings {
-    pub audio: String,
-    pub wait: String,
-    pub fade: String,
-    pub start: String,
-    pub stop: String,
-    pub pause: String,
-    pub load: String,
-    pub group: String,
-}
-
-impl Default for NameFormatSettings {
-    fn default() -> Self {
-        Self {
-            audio: "{filename}".into(),
-            wait: "Wait {duration}".into(),
-            fade: "Fade {targetName}".into(),
-            start: "Start {targetName}".into(),
-            stop: "Stop {targetName}".into(),
-            pause: "Pause {targetName}".into(),
-            load: "Load {targetName}".into(),
-            group: "Group".into(),
-        }
-    }
-}
